Update wheel-encoder.rst
MAVlink2 is required to view the Wheel Distance message node in Mission Planner's MAVlink Inspector. The default setting of the USB port allows the messages, but the default of MAVlink1 on the telemetry port does not. This change reduces confusion when the Wheel Distance is not available over the telemetry port and allows for more effective setup and ground tuning (untethered).